refactor: 불필요 클래스 정리
