+++ src/main/java/nextstep/ladder/domain/edge/LadderPrize.java
@@ -6,4 +6,8 @@ public class LadderPrize {
     public LadderPrize(String value) {
         this.value = value;
     }
+
+    public String getValue() {
+        return value;
+    }
 }
