+++ src/main/java/nextstep/ladder/domain/edge/LadderUser.java
@@ -4,12 +4,18 @@ import java.util.Objects;
 
 public class LadderUser {
     private final String value;
+    private final Position position;
 
     public LadderUser(String name) {
+        this(name, new Position(0));
+    }
+
+    public LadderUser(String name, Position position) {
         if (name.length() > 5) {
             throw new IllegalArgumentException("이름은 최대 5자를 넘을 수 없습니다.");
         }
         this.value = name;
+        this.position = position;
     }
 
     public String getValue() {
