+++ src/main/java/nextstep/ladder/domain/line/Lines.java
@@ -1,9 +1,7 @@
 package nextstep.ladder.domain.line;
 
-import nextstep.ladder.domain.position.Position;
-import nextstep.ladder.domain.position.Positions;
+import nextstep.ladder.domain.edge.Position;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -28,22 +26,22 @@ public class Lines {
         return value.get(0).size();
     }
 
-    public Positions simulate() {
-        Positions res = new Positions();
-        int lineSize = lineSize();
-        for (int start = 0; start <= lineSize; start++) {
-            res.add(simulate(start));
-        }
-        return res;
-    }
-
-    private Position simulate(int start) {
-        Position position = new Position(start);
-        for (Line line: value) {
-            position = line.move(position);
-        }
-       return position;
-    }
+//    public Positions simulate() {
+//        Positions res = new Positions();
+//        int lineSize = lineSize();
+//        for (int start = 0; start <= lineSize; start++) {
+//            res.add(simulate(start));
+//        }
+//        return res;
+//    }
+//
+//    private Position simulate(int start) {
+//        Position position = new Position(start);
+//        for (Line line: value) {
+//            position = line.move(position);
+//        }
+//       return position;
+//    }
 
     public List<List<Boolean>> getvalue() {
         return value.stream().map(Line::getValue).collect(Collectors.toList());
