+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -1,6 +1,5 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.edge.LadderEdges;
 import nextstep.ladder.domain.edge.LadderPrizes;
 import nextstep.ladder.domain.edge.LadderUsers;
 import org.junit.jupiter.api.DisplayName;
@@ -15,8 +14,8 @@ class LadderTest {
     @DisplayName("Ladder 인스턴스 만들기")
     @Test
     public void testConstructor() {
-        LadderUsers ladderUsers =  new LadderUsers(List.of("pobi", "honux", "crong", "jk"));
-        LadderPrizes ladderPrizes = new LadderPrizes(List.of("꽝", "1000", "꽝", "5000"));
-        assertDoesNotThrow(() -> new Ladder(new LadderEdges(ladderUsers, ladderPrizes), 3));
+        LadderUsers users =  new LadderUsers(List.of("pobi", "honux", "crong", "jk"));
+        LadderPrizes prizes = new LadderPrizes(List.of("꽝", "1000", "꽝", "5000"));
+        assertDoesNotThrow(() -> new Ladder(users, prizes, 3));
     }
 }
